Add leaveAdInterestGroup()
As pointed out in #3, there are situations that call for removing a person from the ad targeting pool.

This addition doesn't actually address the use case @rodolpheAV brought up, which is about data mixing between the first-party-and-contextual domain and the interest-group domain; see the issue for more discussion.  But it permits a closely-related flow.