Describe extraSignals feature.
As requested in #119, auctionConfig.extraSignals provides a mechanism
that sellers can use to offer signals to auction participants, similar
to the existing `auctionSignals` and `sellerSignals`, but accepting a
subresource bundle URL instead of a JSON-serializable JavaScript object.

By passing the auction signals this way (and by setting the
X-FLEDGE-Auction-Only header on the response), scripts running on the
publisher page won't be able to read the signals -- only the intended
auction worklets will have access.