Proposal to limit bidding time taken by a buyer
Here's a proposal to add a timeout to limit the entire time each buyer takes to generate bids.  A single timer cover all of a bidder's interest group.  The timer is based on strictly on wall clock time, includes starting processes and downloading resources, and doesn't stop when, e.g., no buyer scripts are running due to CPU contention or whatnot.