replace index 0 of `bidRejectReason`
replace index 0 of `bidRejectReason` from "does not meet k-anonymity threshold" to reject reason not available. When the seller doesn't give a reject reason, buyers will still want to receive the reporting request saying so.
"does not meet k-anonymity threshold" does not seem to be a valid reject reason since not k-anon ads are not checked or rejected by seller script.