Rename dailyUpdateUrl to updateUrl
Updates are currently only done when running an auction (with a minimum of 24 hours between successful updates), not daily in the background, as the name implies.  We have no current plans to change the behavior.

In terms of deprecation approach:  We'll initially accept both, ideally printing a console deprecation warning if dailyUpdateUrl is used when joining an InterestGroup (No warning when the field appears in an update fetch or is used in a bidder script reading its own IG values, though), and we'll copy the field to both dailyUpdateUrl and updateUrl in the interestGroup object we pass to generateBid().  Then at some point before full rollout, we'll remove the old fields.  Previously joined/stored IGs won't need to be updated.