Add ScoringBrowserSignals dictionaryi IDL.
Change |browserSignals| for scoreAd() from an ordered map to an IDL
dictionary {{ScoringBrowserSignals}} so that it can be converted to JS
values easier.

Let methods like [=evaluate a bidding script=] create the realm and pass
it to [=evaluate a script=], so that we can create JS arguments there.

Move realm and agent creation algorithms into a separate section
based on spec review doc's comment.