Remove an obsolete todo about data-version.
The explainer mentioned requesting trusted bidding signals using
multiple requests (for a `generateBid()`), and then need to make
sure that the Data-Version header's value is consistent in these
responses. That's only kept there as an open option (possible
optimization), but not what's implemented. So removing the todo which is
obsolete.