Add Link and Overview on Click/View/Render/Video Interaction Reporting
As an ad tech person reading this page, it would be helpful to get a link from the "reporting" section to more details. I believe those details are in the FF_ads_reporting, which I linked. I also added a bit of color and overview that I think will be helpful to a future ad tech person trying to link these ideas together.