Make Explicit TTL Impact of updateUrl Call
This is one of my questions disguised as a PR that I think would be useful for clarity.

Based on what I'm reading [here](https://wicg.github.io/turtledove/#interest-group-updates), and in the keys list in Section 7, 1.1.12, the TTL is not reset given changes to the IG via the response from updateUrl.

Is that correct? Is that intentional? Why not allow the IG to return `lifetimeMs` and have the TTL reset?