Remove Old Messaging Around Web Bundles and Temporary Nature of CDN Loading
One of the things that came up during TPAC is that the web bundles idea is on hold and that there is no current timeline for them. Since there is not currently a plan to switch to web bundles or trusted CDNs, I wanted to remove the notes about them, as I think they are probably causing more confusion now for someone coming to this and not knowing that history.