Add Some Clarification on Reporting IDs, Origin
Wanted to add some clarity, which also will serve as a confirmation for a question we've had internally, on any limitations on buyer/seller reporting IDs. Added a few others in the neighborhood while at it.

I do want to point out that the buyer/seller reporting IDs are listed in the full spec (https://wicg.github.io/turtledove/) as USVStrings for object declaration, but then DOMString when passed into reporting functions as browser signals. I do see other values go into browser signals as USVStrings, so it seems like either can go in...is that difference intentional?