Remove `reportAdditionalBidResult` for additional bids. (#835)
In the most recent version of the additional bids/negative targeting
design, additional bids are transported via HTTP response headers, so
that sellers don't have a risk for which they'd need a differently named
reporting function to acknowledge that risk. The text in the explainer
correctly noted the risk to buyers, and the differently named reporting
function for buyers, `reportAdditionalBidWin`, remains.