Update guidance on `additionalBidKey` rotation (#841)
* Make negative interest groups non-updatable

In a previous version of the negative targeting design, negative interest groups were updatable so that the additional bid key could be rotated more frequently than every 30 days. We've updated this design so that negative interest groups are non-updatable, and the additional bid key should be rotated exactly every 30 days.

* Make negative interest groups non-updatable

In a previous version of the negative targeting design, negative interest groups were updatable so that the additional bid key could be rotated more frequently than every 30 days. We've updated this design so that negative interest groups are non-updatable, and the additional bid key should be rotated exactly every 30 days.