Emphasize that the additional bid joining origin check fails open (#842)
When an additional bid specifies two or more negative interest groups, it must also identify a single joining origin for all of those negative interest groups. Any negative targeting group joined for a different origin than the one specified is treated as if it's not present. This change emphasizes this policy of "failing open", as the explainer already does for signature verification errors.