Add iframe support for adAuctionHeaders to spec.
This support is provided by a new `adAuctionHeaders` attribute on the
iframe element that will have the same functional behavior as the
`adAuctionHeaders` fetch flag, in that it will trigger the user agent to
send the `Sec-Ad-Auction-Fetch` request header, and to remove
`Ad-Auction-Signals` and `Ad-Auction-Additional-Bid` response headers,
providing their values to the Protected Audience auction.

This is a continuation of https://github.com/WICG/turtledove/pull/883,
which needed to be closed due to an merge conflict on rebase.