Inline the guidance for adAuctionHeaders fetch flag/iframe attribute.
For additional bids, this document previously pointed to the
directFromSellerSignals section for details on how to use the
`adAuctionHeaders` fetch flag or iframe attribute to use response
header interception. To reduce confusion (at the cost of some
repetition), this version copies some of that guidance into the
additional bids section.