Add iframe support for adAuctionHeaders to explainer. (#887)
* Remove `reportAdditionalBidResult` for additional bids.

In the most recent version of the additional bids/negative targeting
design, additional bids are transported via HTTP response headers, so
that sellers don't have a risk for which they'd need a differently named
reporting function to acknowledge that risk. The text in the explainer
correctly noted the risk to buyers, and the differently named reporting
function for buyers, `reportAdditionalBidWin`, remains.

* Add iframe support for adAuctionHeaders to explainer.

This support is provided by a new `adAucitonsHeader` on the iframe attribute that will have the same functional behavior as the `adAuctionHeaders` fetch flag, in that it will trigger the user agent to send the `Sec-Ad-Auction-Fetch` request header, and to remove `Ad-Auction-Signals` and `Ad-Auction-Additional-Bid` response headers, providing their values to the Protected Audience auction.

* Add timing and B&A response blobs for iframe adAuctionHeaders.

Add a few sentences to clarify when Promises should be resolved when
using both fetch and iframe mechanisms for adAuctionHeaders.

Also added a paragraph to the B&A explainer noting the potential use of
iframe navigation to retrieve response blobs.

* Clarify difference between additional bids and directFromSellerSignals.

* Adjust wording for iframe support for Ad-Auction-Result headers.

This is a slight adjustment in wording on the B&A explainer for clarity.

* Clarify guidance on response headers received before runAdAuction.

Specifically, we clarify here that the `directFromSellerSignals` should
be provided as a non-Promise in that scenario, versus as a Promise
that's immediately resolved.

* Inline the guidance for adAuctionHeaders fetch flag/iframe attribute.

For additional bids, this document previously pointed to the
directFromSellerSignals section for details on how to use the
`adAuctionHeaders` fetch flag or iframe attribute to use response
header interception. To reduce confusion (at the cost of some
repetition), this version copies some of that guidance into the
additional bids section.

* Update B&A response blob URL on iframe to match that of fetch request.

* Removing extraneous line left behind from a copy/paste error

* Replace the term 'parameter' with 'field' for auction config fields