Small clarifications / consistency for arbitrary metadata values. (#1054)
* Small clarifications / consistency for arbitrary metadata values.

TechLab standardisation group was discussing recommended serialisation for `perBuyerSignals` from buyers to sellers.  

Suggesting these small cleanups so that it is clear that `perBuyerSignals` is _any_ value (not necessarily a dict or string), and update "arbitrary value" examples to use a uniform `...` placeholder. 


Reference:

Reviewing the spec, `perBuyerSignals` is a map of string keys (buyer origins) to _any _ (arbitrary JSON-serialisable value). Internally, this JS value is serialised to a string: 

```webidl
# Spec:
Promise<record<USVString, any>> perBuyerSignals;

If config["perBuyerSignals"] exists:
1.	Set auctionConfig’s per buyer signals to config["perBuyerSignals"].
2.	Handle an input promise in configuration given auctionConfig and auctionConfig’s per buyer signals:
o	To parse the value result:
1.	Set auctionConfig’s per buyer signals to a new ordered map whose keys are origins and whose values are strings.
2.	For each key → value of result:
1.	Let buyer be the result of parsing an https origin with key. If buyer is failure, throw a TypeError.
2.	Let signalsString be the result of serializing a JavaScript value to a JSON string, given value.
3.	Set auctionConfig’s per buyer signals[buyer] to signalsString.

```

* Cherry pick wording from Stan's arbitrary values PR

* Amend JSON serializable language and 'see also' link to MDN...

As suggested by JensenPaul and caraitto in the comments.