Spec Multibid
This permits generateBid() to return (and setBid to take) a sequence of bids as an alternative to a single one.
A k-anon rerun, still limited to producing a single bid, will now happen when none of the returned bids are k-anonymous (which is equivalent to current behavior when one bid is produced).

To make it easier to produce k-anonymous bids when component ads are involved without relying on a re-run,
this also adds optional targetNumAdComponents and numMandatoryAdComponents fields to bid to request implementations to drop some component ads in order to make a bid k-anonymous.

See #595