Spec: split up response validation into two stages, and fix a bug in it
We will need to do checking in two stages for script bodies to get the
proper cross-site seller signals behavior.

While at it, make sure we actually check that the response is a 2xx;
we are not supposed to be trying to use error pages. Also make the WASM
path go through this, too, so it gets this check as well.