Fix an edge case with join / win history
Due to join / bid history being floored to a multiple of UTC days,
there's a possibility that the interest group outlives its history, even
if it isn't rejoined. This leads to oddities like interest groups with a
join count of 0.

Fix this by letting history live for one more day.