Spec: Avoid overload of contributeToHistogramOnEvent() (#1412)
* Spec: Avoid overload of contributeToHistogramOnEvent()

The Private Aggregation spec now defines this function (for the
aggregate error reporting feature[0]). This has caused the method to be
temporarily overloaded in the Protected Audience spec while the changes
to support this new feature have not yet landed.

To avoid this issue, we modify the handling to use the new spec
mechanism. Note that this PR does not actually implement the new
feature, which will be done in a future PR. This PR should not result in
any functional changes.

[0]: https://github.com/patcg-individual-drafts/private-aggregation-api/blob/main/error_reporting.md

* fix build

* clarify return boolean